janky-thermostat/requirements.txt
simple-pid
dual-mc33926 @ git+https://github.com/Clam-/dual-mc33926@main
pigpio-ads1115 @ git+https://github.com/Clam-/pigpio-ads1115@main
pigpio-sht4x @ git+https://github.com/Clam-/pigpio-sht4x@main
paho-mqtt
pigpio
